# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "https-tunnel control plane",
    "version": "1.0.0",
    "summary": "Issues tunnel sessions and reports on them.",
    "description": "The control plane runs on the base domain, `tunnel.example.com`. It is the\nonly host that serves this API and the web UI.\n\nEvery other host under the base domain, `<label>.tunnel.example.com`, is a\ntunnel. Those hostnames serve whatever the connected client is running locally\nand have no API of their own, so nothing in this document applies to them. A\nrequest to a tunnel hostname whose client is offline is answered by the proxy\nitself with `502` and the offline page.\n\nA client uses two endpoints. `POST /api/v1/connect` registers or resumes a\nsession and returns the public URL, then `GET /api/v1/tunnel` upgrades the\nconnection to the tunnel protocol and keeps it open. The remaining endpoints\nare for inspection and cleanup.\n",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "https-tunnel",
      "url": "/docs"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server, whichever host is serving this document"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
    "/api/v1/auth/account": {
      "post": {
        "operationId": "changeAccount",
        "tags": [
          "auth"
        ],
        "summary": "Change the administrator username, password, or both",
        "description": "There is exactly one administrator account. Sending a new_username renames it rather than creating another, so after admin becomes crwirz nothing signs in as admin again. The password is hashed with PBKDF2-HMAC-SHA256 and written to config.json, and every other signed in browser is dropped.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountChangeRequest"
              },
              "examples": {
                "rename": {
                  "summary": "Rename the account",
                  "value": {
                    "current_password": "…",
                    "new_username": "crwirz"
                  }
                },
                "password": {
                  "summary": "Change the password",
                  "value": {
                    "current_password": "…",
                    "new_password": "a-longer-secret"
                  }
                },
                "both": {
                  "summary": "Both at once, as first boot requires",
                  "value": {
                    "current_password": "admin",
                    "new_username": "crwirz",
                    "new_password": "a-longer-secret"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The account after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            }
          },
          "400": {
            "description": "Nothing to change, or the username or password was rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "auth"
        ],
        "security": [],
        "summary": "Sign in as the administrator",
        "description": "While the server has no password hash the only accepted credentials are admin/admin, and the session it returns can do nothing until the password is changed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The current sign in state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/connect": {
      "post": {
        "operationId": "connect",
        "tags": [
          "tunnel"
        ],
        "summary": "Register or resume a session",
        "description": "Returns the session id and the public URL the tunnel will be served on.\nPassing a `session_id` that the same API key owns resumes that session, keeping its URL; an unknown or foreign id is ignored and a fresh session is issued instead.\nA `subdomain_request` is granted when that label is free, and otherwise silently replaced by a random one, so always read the URL from the response rather than assuming the request was met.\nCalling this repeatedly without a `session_id` creates a new session each time, so clients should persist the id they get back.\n",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRequest"
              },
              "examples": {
                "fresh": {
                  "summary": "Ask for a new session",
                  "value": {}
                },
                "resume": {
                  "summary": "Resume a known session",
                  "value": {
                    "session_id": "6wjulksypp3hbptpm5hqbn2davqbroyn"
                  }
                },
                "request_a_name": {
                  "summary": "Ask for a specific subdomain",
                  "value": {
                    "subdomain_request": "chris"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session issued or resumed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectResponse"
                },
                "example": {
                  "session": "6wjulksypp3hbptpm5hqbn2davqbroyn",
                  "url": "https://qr8xfcb946mw.tunnel.example.com"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "description": "No subdomain could be allocated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "operationId": "listSessions",
        "tags": [
          "sessions"
        ],
        "summary": "List every session",
        "description": "Sessions are returned newest first, connected and disconnected alike.",
        "responses": {
          "200": {
            "description": "The current session table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sessions"
                  ],
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "pruneSessions",
        "tags": [
          "sessions"
        ],
        "summary": "Delete every idle session older than a window",
        "description": "Removes disconnected sessions whose last activity is older than idle_for, freeing their subdomains. Connected tunnels are never removed whatever their age: a tunnel that has been quiet for a week still has a client attached, and cutting it is what deleting a single session by id is for.",
        "parameters": [
          {
            "name": "idle_for",
            "in": "query",
            "required": true,
            "description": "How long a session must have been idle to be removed. The list is fixed so that a mistyped value cannot delete more than intended.",
            "schema": {
              "type": "string",
              "enum": [
                "5m",
                "30m",
                "1h",
                "6h",
                "24h",
                "2d",
                "1w",
                "1m"
              ]
            },
            "example": "24h"
          }
        ],
        "responses": {
          "200": {
            "description": "What was removed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "idle_for",
                    "deleted",
                    "sessions"
                  ],
                  "properties": {
                    "idle_for": {
                      "type": "string"
                    },
                    "deleted": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      },
                      "description": "The sessions that were deleted, oldest first."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "idle_for was missing or not one of the accepted windows.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/v1/sessions/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Session id.",
          "schema": {
            "type": "string"
          },
          "example": "6wjulksypp3hbptpm5hqbn2davqbroyn"
        }
      ],
      "get": {
        "operationId": "getSession",
        "tags": [
          "sessions"
        ],
        "summary": "Fetch one session",
        "responses": {
          "200": {
            "description": "The session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteSession",
        "tags": [
          "sessions"
        ],
        "summary": "Delete a session",
        "description": "Frees the subdomain and disconnects the client if one is attached. The\nclient will reconnect and be issued a new session, on a new URL.\n",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer <api key>`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative to the bearer header; the same key either way."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tunnel_admin",
        "description": "Session cookie set by /api/v1/auth/login. The web UI uses this; API clients use a key."
      }
    },
...
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "AuthStatus": {
        "type": "object",
        "required": [
          "authenticated",
          "must_change_password",
          "password_unset"
        ],
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          },
          "must_change_password": {
            "type": "boolean",
            "description": "True while this session signed in with the default password and has not chosen a new one. Nothing else is readable until it does."
          },
          "password_unset": {
            "type": "boolean",
            "description": "True while the server has no administrator password at all, so the first sign in is admin/admin."
          }
        }
      },
...
      "AccountChangeRequest": {
        "type": "object",
        "required": [
          "current_password"
        ],
        "description": "At least one of new_username and new_password must be present. While the server still has no password of its own, new_password is required whatever else changes.",
        "properties": {
          "current_password": {
            "type": "string",
            "format": "password"
          },
          "new_username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32,
            "pattern": "^[a-z0-9._-]+$",
            "description": "Renames the single administrator account. The previous name stops working immediately, including for the first boot admin/admin path, so it is not left behind as a second way in. Lowercased and trimmed before it is stored and before every comparison."
          },
          "new_password": {
            "type": "string",
            "format": "password",
            "minLength": 8
          }
        }
      }
    }
  }
}